Updated header: new stuff: retroLoadABKSprite, retroPasteSprite, retroBoing
Updated header.
Loading ABK files as sprites: Done
Proto type for PasteSprite, similar to PasteBob, changed the name as bob
object, a sprite is often just graphics, like or not, it what choose.
retroBoing now moved into library, so do that effect with having,
provide that all the program / tests we make.